Link with tensorflow-lite correctly
The aom and aom_static libraries (the encoder side) and the
tf_lite_model and cnn_restore_y4m examples are the only targets that use
tensorflow-lite, so only these targets need to link with
tensorflow-lite.

Also add the -z defs linker option when linking a shared library to
verify that all symbol references in the shared library are resolved.
This part is copied from libaom's CMakeLists.txt.