Fix USE_PER_BLOCK_TMVP to have bit-exact results
The MACRO USE_PER_BLOCK_TMVP is disabled by default, but when enabled, it substitutes the TMVP motion field-building function from a frame-based implementation to a block-based one, providing a reference for interested parties.  Therefore the results from both implementations should be exactly the same.

This MR fixes the current head, so that they do provide exactly the same results.