Bug fix, chroma LR unit size for the bottom row
When calculate the maximum unit size of the bottom low restoration unit, the row offset (8 luma samples) is added. In the current software, the chroma scaling is not taken into account for the row offset in this calculation. This MR fixes this bug.

This bug introduce encoder failure when the image size is around xx.5 multiple of LR unit size.