CWG-D146: move *wt_matrix_ref to CommonQuantParams
After CWG-D146, wt_matrix_ref and iwt_matrix_ref became non-const global
variables. Since both the decoder and encoder initialize wt_matrix_ref
and iwt_matrix_ref during creation, this is not safe if a process
creates a decoder and an encoder or creates multiple decoders or
encoders.

Move wt_matrix_ref and iwt_matrix_ref to the CommonQuantParams struct.
If a process creates only one decoder or only one encoder (the common
case), the memory consumption stays the same.

Ideally, we should allocate the memory for wt_matrix_ref and
iwt_matrix_ref dynamically only when quantization matrices are used.
This could be done as a follow-up. Also, in CWG-D130 we are looking into
using the 8x8, 8x4, and 4x8 QMs directly just like the AV2 spec does.
Then it will be less important to allocate the memory only when QMs are
used.

In test/intrabc_test.cc, allocate xd and cm from the heap because they
are both more than 300 KB. cm grows by 211 KB in this commit.