Avoid extra scale_tx() and calc_wt_matrix() calls
scale_tx() and calc_wt_matrix() only need to be called twice: first for
luma, and then for chroma. If there are two chroma planes (U and V),
don't call extra scale_tx() and calc_wt_matrix() for chroma twice.

Declare scale_tx() and calc_wt_matrix() as static because they are only
used in this file.

In scale_tx(), replace the expressions `c >= 1` with the variable
`plane`.