Fix segment id issue with extended-sdp
In AVM, the mode information for coded block of intra region in inter frames is signaled within function write_mb_modes_kf(), which makes the syntax of segment ID incorrect.

The fix in this MR is putting the signaling of intra region info back to pack_inter_mode_mvs(), which makes the syntax of segment ID consistent between intra region and mixed intra and inter region in inter frames.

Fix issue #545 and #730

No bitstream difference in CTC.