Fix the incorrect motion vector used for boundary detection of current block
When warp is used, `subblock_mv` is set to `mv_refined[ref]`. In the other code path, `subblock_mv` is set to `mv_refined[n_blocks * 2 + ref]`.

The motion vector here is used to measure the amount of the prediction that is off the edge of the screen and is needed when forming the mask to merge the predictions.

This means that the boundary is detected based on the motion vector of the top-left block, not the current block. The fix is use to `n_blocks * 2 + ref` for both places.

Fix issue #416

STATS_CHANGED