Move tcq_mode bit reading to after lossless set up
Prior to this patch, tcq_mode was signaled first and then lossless was deternined per segment, and then if both not coded_lossless and tcq_mode is off, parity hiding could be turned on.

This patch moves tcq_mode setting to after lossless is determined. If a frame is determined to be coded lossless, then tcq_mode is turned off without the need for signaling it.

As before, parity hiding can be on both not lossless and tcq_mode is off.

This is a more consistent design than the prior design where tcq_mode determined lossless, but then lossless determined parity_hiding. Besides, with this patch determination of losslessness is depends only on qindex as in AV1, and not on whether tcq_mode is on or off.

Note: trellis_test was fixed to avoid some sanitizer errors (unrelated issue found by this commit).

STATS_CHANGED