Fix the bug for pred buf reset in RPR case
This fix is in the function av1_compound_single_motion_search() and it solves the rpr-ctc test failures.

This fix does not impact CTC, however it should slightly improve the rpr performance.

In the function av1_compound_single_motion_search(), the following steps are used to set and reset the xd-\>plane\[i\].pre\[idx\] during motion estimation.

**Step1: if RPR, store xd-\>plane\[i\].pre\[idx\] in backup_yv12\[idx\]\[i\]**

**Step 2: if (ref_idx), swap (xd-\>plane\[i\].pre\[0\] and xd-\>plane\[i\].pre\[1\])**

**Step3:  motion estimation process**

**Step4: if RPR, restore xd-\>plane\[i\].pre\[idx\] from backup_yv12\[idx\]\[i\]**

**Step5: if (ref_idx), swap (xd-\>plane\[i\].pre\[0\] and xd-\>plane\[i\].pre\[1\])**

This MR switches the order of step 4 and 5 to have the xd-\>plane\[i\].pre\[idx\] being reset correctly.

STATS_CHANGED