Simplify av1_alloc_qm & av1_free_qm func prototype
Simplify the av1_alloc_qm() and av1_free_qm() function prototypes.
Change av1_alloc_qm() to return the allocated array in the return value
rather than in an output parameter. Change av1_free_qm() to take the
array rather than a pointer to the array as input. These changes remove
one level of indirection.

In the function comment for av1_alloc_qm(), clarify that the flattened
quantization matrix is width-by-height.