Fix on the ref_count value update parts to accurately identify when frame buffers can be released
Problem: The ref_count value was not accurately updated, because the ref_count was updated only when the corresponding frame buffer is added to or removed from a ref_frame_map slot.

Solution: In this MR, the ref_cout value is updated when the corresponding frame buffer is added to or removed from a ref_frame_map slot or the output frame queue (pbi\*\*-\>\*\*output_frames).

In addition, the function check_ref_count_status_enc() is added to check and verify the current status of ref_count values.

Fixes #420