Reduce valid range of QM delta values to -128..127
It is not hard for the encoder to shift the quantization matrix (QM)
coefficient delta values to the range of -128..127 to shorten the svlc()
code, so we should require the smaller valid range.

Remove the symbolic constant NUM_QM_VALS. It actually obscured the code
because we need to know its value to understand the code. Also, since QM
coefficients cannot be zero, strictly speaking the number of QM values
is 255, not 256.

Check for the special value that indicates repeating coefficients after
the % 256 operation.

Issue: https://gitlab.com/AOMediaCodec/avm/-/issues/760