av1_rd_pick_intra_mode_sb: fix for loop to compute `nz`
For 64xN / Nx64 blocks, simply running for loop for tx width \* tx_height is NOT correct, as the actual coeffs will be of size 32xN / Nx32.

This can lead to out-of-bounds access in for loop when trying to read coeffs.

To fix this, we use the helper function to get largest EOB given the transform size.

Fixes #778