Fix for MHCCP reference samples at the tile boundary
Fix #792, the number of MHCCP reference width and height samples in case of tile boudnaries have a "-1" operation, which is desgined for the initial MHCCP design which need "right" and "bottom" samples. After the CWG-E149, the "right" and "bottom" samples are not used anymore, therefore, the "-1" operation can be removed. 

STATS_CHANGED