Guard more code under CONFIG_CORE_TX
This MR extends the coverage of `CONFIG_CORE_TX` to include previously unused code, such as the SIMD implementations of AV1 core transforms. It serves as a preparatory step for a subsequent code cleanup that will remove the `CONFIG_CORE_TX` macro entirely. After this MR is merged, a follow-up MR will be submitted to remove the macro along with all associated disabled code.

Pre-requisite to one item in #764