Update qm_num_repeating_coefs() for CWG-F313
After CWG-F313, the `prev` variable is also updated for the inferred
coefficients above the diagonal in a symmetric matrix. Update
qm_num_repeating_coefs() to return the correct number of repeating
coefficients. Rename it to the more informative name
qm_num_zero_deltas().

Rename the local variables related to the stop symbol.