Turn on F279 by default
Also, the quality of highest level inter frame quality was toned
down a little due to the fact that the TCQ adjustment tool F236
also increased the quality of that same frame. This adjustment
resulted in partial recovery of the overlap gain.

Specifically,
F279 on Avm-v10: -0.15% gain (as proposed);
F279 on baseline after F236: -0.09% gain;
F279 on baseline after F236, readjusted: -0.14% gain (w/ 1 point still running).

Therefore we are proposing the readjusted version in this patch.

STATS_CHANGED