Fix LevelTestLarge.TestTargetLevelApi after !2138
After MR !2138, the maximum number of operating points increased from 32
to 64. So AV1LevelParams::keep_level_stats (which is a packed boolean
bits for all operating points) needs to increase from uint32_t to
uint64_t.

The LevelTestLarge.TestTargetLevelApi test also needs to be adjusted for
the new maximum operating point. The test uses the public API but
MAX_NUM_OPERATING_POINTS is an internal macro, so unfortunately the
maximum operating point is a hardcoded constant in the test.

Fixes https://gitlab.com/AOMediaCodec/avm/-/issues/832.