Fix the interaction between sdp and delta qp
Fix issue #847 

In current AVM code, the delta q parameters are signaled twice, once for luma tree, and once for chroma tree. However, luma tree and chroma tree are sharing the same xd->current_base_qindex. Therefore, the deltaq for luma tree can be overwritten by chroma tree, which is not expected.

The fix is disallowing the signaling of delta qp for chroma qp, which may fix the potential encoder/decoder mismatches if delta qp is used.