CWG-F332: TU64 signaling and decoding is traversed in 128x128 blocks
Fixes #812: TU64×64 are processed in traversal order within each 128×128 unit with this MR. The top-right and bottom-left reference sample availability conditions for intra prediction have also been updated to reflect the modified TU64×64 coding order. This resolves a pipeline issue in CWG-F178 and addresses a cost issue in MV projection based on MV trajectory when the projection window size is 128×128.