Fix on output of trailing frames at the end of coded video sequence (#512)
This MR fixes a bug on outputting remaining showable frames at the end of coded video sequence.

When enable_frame_output_order is equal to 1, the trailing output frames should be output at the end of a coded video sequence.

However, the output operation at the end of CVS was not fully implemented. So, #512 was reported.

This MR fixes #512, by outputting the trailing showable frames when a new CVS starts.