Initialize quantization matrix only when it is used
The computation of QM initialization is significant for decoding time. We initialize it only when it is actually used.

On all intra coding, A3 test set, before the change, decoder profile shows about 5-6% decoding time is spent on QM initialization.

This change avoids the unnecessary computation for most use cases.