Do not initialize wedge masks until the first inter frame
The computation of wedge masks initialization is costly after CWG-F143 and unnecessary for intra frame.

We introduce a flag so that we can move the initialization to the first inter frame.

This significantly improves decoding speed for Allintra mode.

The speed gain varies depending on QP, resolutions, etc. An example is that on some A5 test clips, Allintra mode, QP 210, decoder speed improves up to 30% (125 fps -\> 163 fps).