Fix on output_trailing_frames
This fix will output the frames in the same order as the old code without dropping frames in special cases. This also matches the current specification.

The special case is when the reference with the largest order hint and lowest index is contained multiple times in `cm->ref_frame_map` and there is another reference between the two instances of this frame with a matching order hint.

Further fix for #512