Isolate numLeftSB and numLeftActiveSB. Use numLeftActiveSB for range check and...
In #959, one unwanted change was spotted in intra BC local search range validation. It is not expected to change the colocated buffer range. This fix isolates the SB distance validation (use numLeftActiveSB) and colocated buffer validation (revert to use numLeftSB). When BRU is off or SB size is 128 or 256, numLeftActiveSB and numLeftSB are always the same. It is only when BRU **i**s on && SB size is 64 that numLeftActiveSB and numLeftSB could possibly differ.

Fixes #959