Encoder bugfix of traversed TU64x64 coding order
TU64x64 are coded in Z-order of 128x128 in luma size. However, at encoder side, the processing unit (128x128) was sometimes not adjusted based on chroma format, and this MR fixes this by including horizontal and vertical sampling size when determining the processing unit size of TU64x64. The fix is encoder only, and resolves issue #936.