Fix consistency in refined MV reused for chroma
When DAMR was removed in !1586, some code for chroma reuse logic of luma refined MVs was removed, which created an inconsistency issue that luma refined MV are still reused for TIP blocks but not reused for non-TIP blocks. To fix this issue, this patch brings back the reuse logic for non-TIP, as was intended in CWG-F064.

Other relevant minor fixes are included:
- Fix subblock offsets in chroma for frame boundary check. It has no impact without the reuse was introduced.
- Fix a condition for MV averaging when a 4x4 chroma block has multiple colocated luma refined MVs. Previously this averaging was mistakenly performed for 8x16 luma blocks that do not need it. This change has noise level performance impact.
- Fix a compile warning and some cleanups

Fixes #956 

STATS_CHANGED